Refactor CLI commands to use shared SDK operations
