+++ mcpjam-inspector/server/utils/route-handlers.ts
@@ -1,121 +1,39 @@
 /**
  * Shared route handler functions for resources, prompts, and tools.
  *
- * Pure functions: (manager, params) → result.
- * No Hono context, no c.json(), no error formatting.
+ * Pure operations re-exported from @mcpjam/sdk, plus inspector-specific
+ * listTools enrichment (toolsMetadata + tokenCount).
  *
- * Used by both web/ and mcp/ route sets, following the prepareChatV2() pattern.
+ * Used by both web/ and mcp/ route sets.
  */
 
+// Re-export pure operations from SDK — no behavioral changes
+export {
+  listResources,
+  readResource,
+  listPrompts,
+  listPromptsMulti,
+  getPrompt,
+} from "@mcpjam/sdk";
+
 import type { MCPClientManager } from "@mcpjam/sdk";
+import { listTools as listToolsBase } from "@mcpjam/sdk";
 import { countToolsTokens } from "./tokenizer-helpers.js";
 
 type Manager = InstanceType<typeof MCPClientManager>;
 
-// ── Resources ────────────────────────────────────────────────────────
-
-export async function listResources(
-  manager: Manager,
-  params: { serverId: string; cursor?: string },
-) {
-  const result = await manager.listResources(
-    params.serverId,
-    params.cursor ? { cursor: params.cursor } : undefined,
-  );
-  return {
-    resources: result.resources ?? [],
-    nextCursor: result.nextCursor,
-  };
-}
-
-export async function readResource(
-  manager: Manager,
-  params: { serverId: string; uri: string },
-) {
-  const content = await manager.readResource(params.serverId, {
-    uri: params.uri,
-  });
-  return { content };
-}
-
-// ── Prompts ──────────────────────────────────────────────────────────
-
-export async function listPrompts(
-  manager: Manager,
-  params: { serverId: string; cursor?: string },
-) {
-  const result = await manager.listPrompts(
-    params.serverId,
-    params.cursor ? { cursor: params.cursor } : undefined,
-  );
-  return {
-    prompts: result.prompts ?? [],
-    nextCursor: result.nextCursor,
-  };
-}
-
-export async function listPromptsMulti(
-  manager: Manager,
-  params: { serverIds: string[] },
-) {
-  const promptsByServer: Record<string, unknown[]> = {};
-  const errors: Record<string, string> = {};
-
-  await Promise.all(
-    params.serverIds.map(async (serverId) => {
-      try {
-        const { prompts } = await manager.listPrompts(serverId);
-        promptsByServer[serverId] = prompts ?? [];
-      } catch (error) {
-        const errorMessage =
-          error instanceof Error ? error.message : "Unknown error";
-        errors[serverId] = errorMessage;
-        promptsByServer[serverId] = [];
-      }
-    }),
-  );
-
-  const payload: Record<string, unknown> = { prompts: promptsByServer };
-  if (Object.keys(errors).length > 0) {
-    payload.errors = errors;
-  }
-  return payload;
-}
-
-export async function getPrompt(
-  manager: Manager,
-  params: {
-    serverId: string;
-    name: string;
-    arguments?: Record<string, unknown>;
-  },
-) {
-  const promptArguments = params.arguments
-    ? Object.fromEntries(
-        Object.entries(params.arguments).map(([key, value]) => [
-          key,
-          String(value),
-        ]),
-      )
-    : undefined;
-
-  const content = await manager.getPrompt(params.serverId, {
-    name: params.name,
-    arguments: promptArguments,
-  });
-  return { content };
-}
-
-// ── Tools ────────────────────────────────────────────────────────────
-
+/**
+ * Inspector-enriched listTools: adds toolsMetadata and optional tokenCount
+ * on top of the SDK's pure listTools.
+ */
 export async function listTools(
   manager: Manager,
   params: { serverId: string; modelId?: string; cursor?: string },
 ) {
-  const result = await manager.listTools(
-    params.serverId,
-    params.cursor ? { cursor: params.cursor } : undefined,
-  );
+  const result = await listToolsBase(manager, {
+    serverId: params.serverId,
+    cursor: params.cursor,
+  });
 
   const toolsMetadata = manager.getAllToolsMetadata(params.serverId);
 
@@ -127,6 +45,5 @@ export async function listTools(
     ...result,
     toolsMetadata,
     tokenCount,
-    nextCursor: result.nextCursor,
   };
 }
